Fix types in Stats and Status structs
Most of these fields should be unsigned integers.
I’ve also replaced the numbers that represent durations with
time::Duration because that’s what a consumer of the library would
expect to get there.
The time crate was choosen because it provides serde integration and
implements Default for its types which minimizes the necessary changes
to existing code.